multer added in deleteBook controller

+++ controllers/book.js
@@ -1,4 +1,5 @@
 const Book = require("../models/book");
+const fs = require("fs");
 
 exports.createBook = (req, res, next) => {
   //Potentiel erreur : parse(req.body)
@@ -73,7 +74,28 @@ exports.getAllBooks = (req, res, next) => {
     .catch((error) => res.status(400).json({ error }));
 };
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Livre supprimé !" }))
-    .catch((error) => res.status(400).json({ error }));
+  Book.findOne({ _id: req.params.id })
+    .then((book) => {
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: "Not authorized" });
+      } else {
+        const filename = book.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Book.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: "Objet supprimé !" });
+            })
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch((error) => {
+      res.status(500).json({ error });
+    });
 };
+// Delete book originel ->
+// exports.deleteBook = (req, res, next) => {
+//   Book.deleteOne({ _id: req.params.id })
+//     .then(() => res.status(200).json({ message: "Livre supprimé !" }))
+//     .catch((error) => res.status(400).json({ error }));
+// };
